PDF-HUL-117: rotation property not a Simple object
Hi @carlwilson,
I was working on the error PDF-HUL-117 and I tested the provided Example. I noticed the rotation property was null and that this should process properly. The error what occurred was a casting error. The problem is that the object returned for DICT_KEY_ROTATE was a PdfIndirectObj and not a PdfSimpleObject. So I added code to make the PdfIndirectObj a PdfSimpleObject. Now the Example is processed without any problems and is the file valid and well-formed.
I expect the same thing can happen for the other properties. If you approve this, I change the other properties as well.

Sam